+++ src/Packages/Insight/Dataflow/Modules/Filters/NeighborhoodConnectedImageFilter.cc
@@ -151,7 +151,7 @@ NeighborhoodConnectedImageFilter::run( itk::Object *obj_InputImage)
   // clear defined object guis if things aren't in sync
   
   
-  if(radius.GetSizeDimension() != gui_dimension_.get()) { 
+  if((int)radius.GetSizeDimension() != gui_dimension_.get()) { 
     gui_dimension_.set(radius.GetSizeDimension());    
   
     // for each defined object, clear gui
@@ -181,7 +181,7 @@ NeighborhoodConnectedImageFilter::run( itk::Object *obj_InputImage)
   }
 
   // set radius values
-  for(int i=0; i<radius.GetSizeDimension(); i++) {
+  for(unsigned int i=0; i<radius.GetSizeDimension(); i++) {
     radius[i] = gui_radius_[i]->get();
   }
 
@@ -204,7 +204,7 @@ NeighborhoodConnectedImageFilter::run( itk::Object *obj_InputImage)
   }
 
   // set seed_point values
-  for(int i=0; i<seed_point.GetIndexDimension(); i++) {
+  for(unsigned int i=0; i<seed_point.GetIndexDimension(); i++) {
     seed_point[i] = gui_seed_point_[i]->get();
   }
